requirements.txt
charset-normalizer==3.4.4
pillow==12.0.0
pyStrich==0.10
reportlab==4.4.4
